added: colored help pages and new shorthand commands
