+++ compiler/src/onyx.c
@@ -30,405 +30,9 @@ extern struct bh_allocator global_heap_allocator;
     #define ONYX_RUNTIME_LIBRARY_MAPPED none
 #endif
 
-#define STRINGIFY_(x) #x
-#define STRINGIFY(x) STRINGIFY_(x)
-
-
 Context context;
 
-#define VERSION_STRING "Onyx toolchain version " VERSION "\n" \
-    "Built on " __TIMESTAMP__ "\n" \
-    "Runtime: " STRINGIFY(ONYX_RUNTIME_LIBRARY_MAPPED) "\n"
-
-#define DOCSTRING_HEADER VERSION_STRING \
-    "\n" \
-    "The toolchain for the Onyx programming language, created by Brendan Hansen.\n" \
-    "\n"
-
-
-static const char* top_level_docstring = DOCSTRING_HEADER
-    "Usage:\n"
-    "\tonyx <subcommand>\n"
-    "\n"
-    "Subcommands:\n"
-    "\thelp      Shows this help message. Use \"onyx help <subcommand>\".\n"
-    "\tbuild     Compiles an Onyx program into an executable.\n"
-#ifdef ONYX_RUNTIME_LIBRARY
-    "\trun       Compiles and runs an Onyx program, all at once.\n"
-#endif
-    "\tcheck     Checks syntax and types of an Onyx program.\n"
-#ifdef _BH_LINUX
-    "\twatch     Continuously rebuilds an Onyx program on file changes.\n"
-#endif
-    "\tpackage   Package manager\n"
-    "\tversion   Prints version information\n";
-    // "\tdoc <input files>\n"
-
-static const char *build_docstring = DOCSTRING_HEADER
-    "Usage:\n"
-    "\tonyx %s <input files> [-o target_file] OPTIONS\n"
-    "\n"
-    "Required:\n"
-    "\t<input files>           One or more Onyx files to include in the program.\n"
-    "\n"
-    "Options:\n"
-    "\t-o <target_file>        Specify the target file (default: out.wasm).\n"
-    "\t   --output <target_file>\n"
-    "\t-I <dir>                Include a directory in the search path.\n"
-    "\t--runtime, -r <runtime> Specifies the runtime. Can be: onyx, wasi, js, custom.\n"
-    "\t                        (default: onyx)\n"
-    "\t--verbose, -V           Verbose output.\n"
-    "\t           -VV          Very verbose output.\n"
-    "\t           -VVV         Very very verbose output (to be used by compiler developers).\n"
-    "\t--multi-threaded        Enables multi-threading for this compilation.\n"
-    "\t                        Automatically enabled for \"onyx\" runtime.\n"
-    "\t--doc <doc_file>        Generates an O-DOC file, a.k.a an Onyx documentation file. Used by onyx-doc-gen.\n"
-    "\t--tag                   Generates a C-Tag file.\n"
-    "\t--syminfo <target_file> (DEPRECATED) Generates a symbol resolution information file. Used by onyx-lsp.\n"
-    "\t--lspinfo <target_file> Generates an LSP information file. Used by onyx-lsp.\n"
-    "\t--stack-trace           Enable dynamic stack trace.\n"
-    "\t--no-core               Disable automatically including \"core/module\".\n"
-    "\t--no-stale-code         Disables use of `#allow_stale_code` directive\n"
-    "\t--no-type-info          Disables generating type information\n"
-    "\t--generate-method-info  Populate method information in type information structures.\n"
-    "\t                        Can drastically increase binary size.\n"
-    "\t--generate-foreign-info Generate information for foreign blocks. Rarely needed, so disabled by default.\n"
-    "\t--wasm-mvp              Use only WebAssembly MVP features.\n"
-    "\t--feature <feature>     Enable an experimental language feature.\n"
-    "\n"
-    "Developer options:\n"
-    "\t--no-colors               Disables colors in the error message.\n"
-    "\t--no-file-contents        Disables '#file_contents' for security.\n"
-    "\t--error-format (v1|v2)    Changes the output error format.\n"
-    "\t--show-all-errors         Print all errors (can result in many consequencial errors from a single error)\n"
-    "\t--print-function-mappings Prints a mapping from WASM function index to source location.\n"
-    "\t--print-static-if-results Prints the conditional result of each #if statement. Useful for debugging.\n"
-    "\n";
-
-
-static CompileOptions compile_opts_parse(bh_allocator alloc, int argc, char *argv[]) {
-    CompileOptions options = {
-        .allocator = alloc,
-        .action = ONYX_COMPILE_ACTION_PRINT_HELP,
-
-        .verbose_output          = 0,
-        .fun_output              = 0,
-        .print_function_mappings = 0,
-        .no_file_contents        = 0,
-
-        .use_post_mvp_features   = 1,
-        .use_multi_threading     = 0,
-        .generate_foreign_info   = 0,
-        .generate_type_info      = 1,
-        .generate_method_info    = 0,
-        .no_core                 = 0,
-        .no_stale_code           = 0,
-        .show_all_errors         = 0,
-
-        .enable_optional_semicolons = 0,
-
-        .runtime = Runtime_Onyx,
-
-        .files = NULL,
-        .target_file = "out.wasm",
-
-        .documentation_file = NULL,
-        .symbol_info_file   = NULL,
-        .help_subcommand    = NULL,
-
-        .defined_variables = NULL,
-
-        .debug_info_enabled = 0,
-
-        .passthrough_argument_count = 0,
-        .passthrough_argument_data  = NULL,
-
-        .generate_tag_file = 0,
-        .generate_symbol_info_file = 0,
-        .generate_lsp_info_file = 0,
-
-        .running_perf = 0,
-
-        .error_format = "v1",
-    };
-
-    bh_arr_new(alloc, options.files, 2);
-    bh_arr_new(alloc, options.included_folders, 2);
-    bh_arr_new(alloc, options.defined_variables, 2);
-
-    char* core_installation = NULL;
-
-    #if defined(_BH_LINUX) || defined(_BH_DARWIN)
-    core_installation = getenv("ONYX_PATH");
-
-    if (getenv("ONYX_ERROR_FORMAT")) {
-        options.error_format = getenv("ONYX_ERROR_FORMAT");
-    }
-    #endif
-    #ifdef _BH_WINDOWS
-    char *tmp_core_installation = bh_alloc_array(alloc, u8, 512);
-    char *tmp_error_format      = bh_alloc_array(alloc, u8, 512);
-
-    if (GetEnvironmentVariableA("ONYX_PATH", tmp_core_installation, 512) > 0) {
-        core_installation = tmp_core_installation;
-    }
-    if (GetEnvironmentVariableA("ONYX_ERROR_FORMAT", tmp_error_format, 512) > 0) {
-        options.error_format = tmp_error_format;
-    }
-    #endif
-
-    if (core_installation == NULL) {
-        bh_printf("error: ONYX_PATH environment variable is not set. Please set this to the location of your Onyx installation.\n");
-        exit(1);
-    }
-
-    // NOTE: Add the current folder
-    bh_arr_push(options.included_folders, core_installation);
-    bh_arr_push(options.included_folders, ".");
-
-    if (argc == 1) return options;
-    i32 arg_parse_start = 1;
-
-    if (!strcmp(argv[1], "help")) {
-        options.action = ONYX_COMPILE_ACTION_PRINT_HELP;
-        options.help_subcommand = argc > 2 ? argv[2] : NULL;
-    }
-    else if (!strcmp(argv[1], "version")) {
-        options.action = ONYX_COMPILE_ACTION_PRINT_VERSION;
-        goto skip_parsing_arguments;
-    }
-    else if (!strcmp(argv[1], "compile") || !strcmp(argv[1], "build")) {
-        options.action = ONYX_COMPILE_ACTION_COMPILE;
-        arg_parse_start = 2;
-    }
-    else if (!strcmp(argv[1], "check")) {
-        options.action = ONYX_COMPILE_ACTION_CHECK;
-        arg_parse_start = 2;
-    }
-    else if (!strcmp(argv[1], "pkg") || !strcmp(argv[1], "package")) {
-        options.action = ONYX_COMPILE_ACTION_RUN;
-        options.passthrough_argument_count = argc - 2;
-        options.passthrough_argument_data  = &argv[2];
-        options.generate_method_info = 1; // The package manager needs this to be enabled.
-        arg_parse_start = argc;
-
-        bh_arr_push(options.files, bh_aprintf(alloc, "%s/tools/onyx-pkg.onyx", core_installation));
-        goto skip_parsing_arguments;
-    }
-    #ifdef ONYX_RUNTIME_LIBRARY
-    else if (!strcmp(argv[1], "run")) {
-        options.action = ONYX_COMPILE_ACTION_RUN;
-        arg_parse_start = 2;
-    }
-    #endif
-    #ifdef _BH_LINUX
-    else if (!strcmp(argv[1], "watch")) {
-        options.action = ONYX_COMPILE_ACTION_WATCH;
-        arg_parse_start = 2;
-    }
-        // `#ifdef ONYX_RUNTIME_LIBRARY
-        // `else if (!strcmp(argv[1], "run-watch")) {
-        // `    options.action = ONYX_COMPILE_ACTION_RUN_WATCH;
-        // `    arg_parse_start = 2;
-        // `}
-        // `#endif
-    #endif
-    else {
-        char *script_filename = bh_aprintf(alloc, "%s/tools/%s.wasm", core_installation, argv[1]);
-        if (bh_file_exists(script_filename)) {
-            options.action = ONYX_COMPILE_ACTION_RUN_WASM;
-            options.target_file = script_filename;
-
-            options.passthrough_argument_count = argc - 2;
-            options.passthrough_argument_data  = &argv[2];
-            goto skip_parsing_arguments;
-        }
-
-        bh_printf("Unknown subcommand: '%s'\n", argv[1]);
-        bh_printf("Run \"onyx help\" for valid subcommands.\n");
-        exit(1);
-    }
-
-    if (options.action != ONYX_COMPILE_ACTION_PRINT_HELP) {
-        fori(i, arg_parse_start, argc) {
-            if (!strcmp(argv[i], "-o") || !strcmp(argv[i], "--output")) {
-                options.target_file = argv[++i];
-            }
-            else if (!strcmp(argv[i], "--verbose") || !strcmp(argv[i], "-V")) {
-                options.verbose_output = 1;
-            }
-            else if (!strcmp(argv[i], "-VV")) {
-                options.verbose_output = 2;
-            }
-            else if (!strcmp(argv[i], "-VVV")) {
-                options.verbose_output = 3;
-            }
-            else if (!strcmp(argv[i], "--print-function-mappings")) {
-                options.print_function_mappings = 1;
-            }
-            else if (!strcmp(argv[i], "--print-static-if-results")) {
-                options.print_static_if_results = 1;
-            }
-            else if (!strcmp(argv[i], "--no-colors")) {
-                options.no_colors = 1;
-            }
-            else if (!strcmp(argv[i], "--no-file-contents")) {
-                options.no_file_contents = 1;
-            }
-            else if (!strcmp(argv[i], "--wasm-mvp")) {
-                options.use_post_mvp_features = 0;
-            }
-            else if (!strcmp(argv[i], "--multi-threaded")) {
-                options.use_multi_threading = 1;
-            }
-            else if (!strcmp(argv[i], "--generate-foreign-info")) {
-                options.generate_foreign_info = 1;
-            }
-            else if (!strcmp(argv[i], "--generate-method-info")) {
-                options.generate_method_info = 1;
-            }
-            else if (!strcmp(argv[i], "--no-type-info")) {
-                options.generate_type_info = 0;
-            }
-            else if (!strcmp(argv[i], "--no-core")) {
-                options.no_core = 1;
-            }
-            else if (!strcmp(argv[i], "--no-stale-code")) {
-                options.no_stale_code = 1;
-            }
-            else if (!strcmp(argv[i], "--show-all-errors")) {
-                options.show_all_errors = 1;
-            }
-            else if (!strcmp(argv[i], "--error-format")) {
-                options.error_format = argv[++i];
-            }
-            else if (!strcmp(argv[i], "--feature")) {
-                char *next_arg = argv[++i];
-                if (!strcmp(next_arg, "optional-semicolons")) {
-                    options.enable_optional_semicolons = 1;
-                }
-            }
-            else if (!strcmp(argv[i], "-I")) {
-                bh_arr_push(options.included_folders, argv[++i]);
-            }
-            else if (!strncmp(argv[i], "-D", 2)) {
-                i32 len = strlen(argv[i]);
-                i32 j=2;
-                while (argv[i][j] != '=' && j < len) j++;
-
-                if (j < len) argv[i][j] = '\0';
-
-                DefinedVariable dv;
-                dv.key   = argv[i] + 2;
-                dv.value = argv[i] + j + 1;
-                bh_arr_push(options.defined_variables, dv);
-            }
-            else if (!strcmp(argv[i], "-r") || !strcmp(argv[i], "--runtime")) {
-                i += 1;
-                if      (!strcmp(argv[i], "onyx"))   options.runtime = Runtime_Onyx;
-                else if (!strcmp(argv[i], "wasi"))   options.runtime = Runtime_Wasi;
-                else if (!strcmp(argv[i], "js"))     options.runtime = Runtime_Js;
-                else if (!strcmp(argv[i], "custom")) options.runtime = Runtime_Custom;
-                else {
-                    bh_printf("WARNING: '%s' is not a valid runtime. Defaulting to 'onyx'.\n", argv[i]);
-                    options.runtime = Runtime_Onyx;
-                }
-            }
-            else if (!strcmp(argv[i], "--doc")) {
-                options.documentation_file = argv[++i];
-            }
-            else if (!strcmp(argv[i], "--tag")) {
-                options.generate_tag_file = 1;
-            }
-            else if (!strcmp(argv[i], "--syminfo")) {
-                options.generate_symbol_info_file = 1;
-                options.symbol_info_file = argv[++i];
-            }
-            else if (!strcmp(argv[i], "--lspinfo")) {
-                options.generate_symbol_info_file = 1;
-                options.generate_lsp_info_file = 1;
-                options.symbol_info_file = argv[++i];
-            }
-            else if (!strcmp(argv[i], "--debug")) {
-                options.debug_session = 1;
-                options.debug_info_enabled = 1;
-                options.stack_trace_enabled = 1;
-            }
-            else if (!strcmp(argv[i], "--debug-info")) {
-                options.debug_info_enabled = 1;
-                options.stack_trace_enabled = 1;
-            }
-            else if (!strcmp(argv[i], "--stack-trace")) {
-                options.stack_trace_enabled = 1;
-            }
-            else if (!strcmp(argv[i], "--perf")) {
-                options.running_perf = 1;
-            }
-            else if (!strcmp(argv[i], "--")) {
-                options.passthrough_argument_count = argc - i - 1;
-                options.passthrough_argument_data  = &argv[i + 1];
-                break;
-            }
-#if defined(_BH_LINUX)
-            // NOTE: Fun output is only enabled for Linux because Windows command line
-            // is not ANSI compatible and for a silly feature, I don't want to learn
-            // how to properly do arbitrary graphics in it.
-            else if (!strcmp(argv[i], "--fun") || !strcmp(argv[i], "-F")) {
-                options.fun_output = 1;
-            }
-#endif
-            else {
-                if (bh_str_ends_with(argv[i], ".wasm") && options.action == ONYX_COMPILE_ACTION_RUN) {
-                    if (bh_arr_length(options.files) > 0) {
-                        bh_printf("Expected only one '.wasm', or multiple '.onyx' files to be given, not a mixture.\n");
-                        exit(1);
-                    }
-
-                    options.action = ONYX_COMPILE_ACTION_RUN_WASM;
-                    options.target_file = argv[i];
-
-                    options.passthrough_argument_count = argc - i - 1;
-                    options.passthrough_argument_data  = &argv[i + 1];
-                    break;
-
-                } else {
-                    bh_arr_push(options.files, argv[i]);
-                }
-            }
-        }
-    }
-
-  skip_parsing_arguments:
-
-    // NOTE: Always enable multi-threading for the Onyx runtime.
-    if (options.runtime == Runtime_Onyx) {
-        options.use_multi_threading = 1;
-    }
-
-    return options;
-}
-
-static void compile_opts_free(CompileOptions* opts) {
-    bh_arr_free(opts->files);
-    bh_arr_free(opts->included_folders);
-}
-
-static void print_subcommand_help(const char *subcommand) {
-    if (!strcmp(subcommand, "build")
-        || !strcmp(subcommand, "run")
-        || !strcmp(subcommand, "check")
-        || !strcmp(subcommand, "watch")) {
-        bh_printf(build_docstring, subcommand);
-    }
-
-    else  {
-        bh_printf("Unknown subcommand: '%s'\n", subcommand);
-        bh_printf("Run \"onyx help\" for valid subcommands.\n");
-        exit(1);
-    }
-}
-
-
+#include "./cli.c"
 
 typedef enum CompilerProgress {
     ONYX_COMPILER_PROGRESS_ERROR,
@@ -1089,16 +693,6 @@ static CompilerProgress onyx_flush_module() {
     if (context.options->verbose_output)
         bh_printf("Outputting to WASM file:   %s\n", output_file.filename);
 
-    // APPARENTLY... the WebAssembly Threading proposal says that the data segment initializations
-    // in a WASM module are copied into the linear memory EVERY time the module is instantiated, not
-    // just the first time. This means that if we are happily chugging along and modifying global state
-    // and then we spawn a thread, that thread will completely wipe all changes to the global and return
-    // it to its original state. This is horrible obviously, but the only thing that is more horrible is
-    // that the best way around this is to create a second WASM module that simply initializes the given
-    // data section. Then have a section module that is actually your code. For right now, this is going
-    // to be fine since the browser is really the only place that multi-threading can be used to any
-    // degree of competency. But still... This is god awful and I hope that there is some other way to
-    // around this down the line.
     if (context.options->use_multi_threading && !context.options->use_post_mvp_features) {
         bh_file data_file;
         if (bh_file_create(&data_file, bh_aprintf(global_scratch_allocator, "%s.data", context.options->target_file)) != BH_FILE_ERROR_NONE)
@@ -1295,9 +889,7 @@ int main(int argc, char *argv[]) {
                 }
             }
             break;
-        #endif
 
-        #ifdef ONYX_RUNTIME_LIBRARY
         case ONYX_COMPILE_ACTION_RUN_WASM:
             global_heap_allocator = bh_heap_allocator();
             context_init(&compile_opts);
@@ -1305,6 +897,7 @@ int main(int argc, char *argv[]) {
             if (!onyx_run_wasm_file(context.options->target_file)) {
                 compiler_progress = ONYX_COMPILER_PROGRESS_ERROR;
             }
+            break;
         #endif
 
         default: break;
